+++ token-repository/src/test/scala/io/renku/tokenrepository/repository/AccessTokenCryptoSpec.scala
@@ -21,7 +21,7 @@ package io.renku.tokenrepository.repository
 import cats.syntax.all._
 import com.typesafe.config.ConfigFactory
 import io.renku.config.ConfigLoader.ConfigLoadingException
-import io.renku.crypto.AesCrypto.Secret
+import io.renku.crypto.Secret
 import io.renku.generators.CommonGraphGenerators._
 import io.renku.generators.Generators.Implicits._
 import io.renku.generators.Generators._
